who needs a global variable
so the eg from the book actually used a global variable to set the starting number and then worked from there, whereas i somehow made it work without a global. i do not know if this is a good or bad thing. also just used the same awful css from the previous activity because 0 fucks